move collect fields to utils
