+++ .changeset/chatty-worms-pull.md
@@ -1,0 +1,5 @@
+---
+'@graphql-tools/utils': minor
+---
+
+add `@defer` directive
