+++ .changeset/spotty-suns-bake.md
@@ -1,0 +1,5 @@
+---
+'@graphql-tools/utils': minor
+---
+
+add `@stream` directive
