+++ .changeset/warm-bags-camp.md
@@ -1,0 +1,7 @@
+---
+'@graphql-tools/utils': major
+'@graphql-tools/delegate': patch
+'@graphql-tools/stitch': patch
+---
+
+update `collectFields` to support collecting deffered values
