--- packages/executor/src/directives/index.ts
@@ -1,2 +1,0 @@
-export * from './defer.js';
-export * from './stream.js';
