+++ packages/executor/src/index.ts
@@ -1,2 +1,1 @@
 export * from './execution/index.js';
-export * from './directives/index.js';
