+++ packages/utils/src/directives.ts
@@ -8,6 +8,26 @@ import {
 } from 'graphql';
 
 /**
+ * Used to conditionally defer fragments.
+ */
+export const GraphQLDeferDirective = new GraphQLDirective({
+  name: 'defer',
+  description: 'Directs the executor to defer this fragment when the `if` argument is true or undefined.',
+  locations: [DirectiveLocation.FRAGMENT_SPREAD, DirectiveLocation.INLINE_FRAGMENT],
+  args: {
+    if: {
+      type: new GraphQLNonNull(GraphQLBoolean),
+      description: 'Deferred when true or undefined.',
+      defaultValue: true,
+    },
+    label: {
+      type: GraphQLString,
+      description: 'Unique name',
+    },
+  },
+});
+
+/**
  * Used to conditionally stream list fields.
  */
 export const GraphQLStreamDirective = new GraphQLDirective({
