+++ packages/utils/src/index.ts
@@ -52,3 +52,4 @@ export * from './getOperationASTFromRequest.js';
 export * from './extractExtensionsFromSchema.js';
 export * from './Path.js';
 export * from './jsutils.js';
+export * from './directives.js';
